Very basic http -> microservice example
This constellation enables the most primitive communication, where the HTTP service can send a message to the microservice and get a response back.

The HTTP service has no microservice capabilities on its own in this setup, so it is no longer a hybrid application at this point in time (i.e. it doesn't call `connectMicroservice`). But that'll be easy to re-introduce when necessary.